requirements.txt
# ================================================================================
# TOPIK DAILY — Python Dependencies
# ================================================================================

# Core
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
aiohttp>=3.9.0

# Google APIs
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# Audio/Video Processing
pydub>=0.25.1
mutagen>=1.47.0
azure-cognitiveservices-speech>=1.32.0
edge-tts>=6.1.9

# Document Generation
python-docx>=1.0.0
Pillow>=10.1.0
reportlab>=4.0.0

# Web Scraping
selenium>=4.15.0
webdriver-manager>=4.0.1
lxml>=4.9.0

# RSS/XML
feedparser>=6.0.10

# Telegram Bot
python-telegram-bot>=20.7

# Anki Flashcards
genanki>=0.13.0

# Utilities
typing-extensions>=4.8.0
